Fix tests: handle change in error message & node 18 fetch issue
It appears some dependency update recently has changed the IPFS error
message when connection are reset to be more explicit. This is good
really, but breaks out tests that check the message, so this commit
updates that.

Also, in Node 18, fetch is available by default (though experimental)
backed by Undici. Unfortunately, it's buggy, and fails to handle
responses with no transfer-encoding or content-length headers. We work
around this by setting an explicit transfer encoding header for now.